feat: add new rotate_selected mode (#171)
adds the `rotationResolution` configuration to control how much mouse movement affects the ROI's angle in this mode.

BREAKING-CHANGE: rename `allowRotate` to `displayRotationHandle` on `RoiList`. The old option did not have a good name because it was only controlling the ability to rotate by showing and hiding the handle. There are other ways now to rotate an ROI.